fix(ruff): 启用 ASYNC/DTZ/RUF 规则族并清理存量（OSS E1，P1）
- pyproject: select 增 ASYNC/DTZ/RUF；ignore 增 RUF001-003/ASYNC109（报告建议直接 ignore）
  + RUF102（误伤 check_redlines 自定义红线 noqa 协议 # noqa: R16）
- RUF100/RUF102 死 noqa 自动清理；RUF005/022/023/RUF015/RUF046/RUF034 风格项修复
- RUF012（19）类属性可变默认值全部 ClassVar 化
- DTZ（43）：DTZ005 now()→get_now()、DTZ011 today()→get_now().date()（统一时区入口）；
  DTZ007/901 业务日期字符串/哨兵边界加 noqa+理由（R3）；RUF009 dataclass 默认值改 default_factory
- ASYNC240 生产 11 处逐条判定加 noqa（低危纯路径操作，报告推荐做法）
- per-file-ignores：tests//scripts/ 豁免测试常态规则，生产代码严格全开（0 命中）
- 恢复 RUF100 --fix 误删的 8 处 VM # noqa: R16 豁免标记（幂等单例持有引用，R16 存量豁免）
- 全量 ruff All checks passed；ruff format 通过；check_redlines 0 errors
- 关键相关单测：time_utils/scheduler/task_manager/diagnostics 352 passed；backtest 21 passed

+++ ui/components/slider_input.py
@@ -87,7 +87,7 @@ def SliderInput(
     ft.use_effect(lambda: set_draft(formatter(value)), [value])
 
     if divisions is None:
-        divisions = int(round((max_val - min_val) / step)) if step > 0 else None
+        divisions = round((max_val - min_val) / step) if step > 0 else None  # RUF046: round 已返回 int，无需再 int()
 
     def _on_slider_change(e: ft.ControlEvent) -> None:
         raw = get_control_value(e.control, ft.Slider) if e and e.control else value
